[FIX] reorganize api call files structure
